+++ src/explorepy/explore.py
@@ -19,6 +19,7 @@ from appdirs import user_cache_dir
 from threading import Timer
 import logging
 import numpy as np
+import re
 
 import explorepy
 from explorepy.tools import create_exg_recorder, create_orn_recorder, create_marker_recorder, LslServer, PhysicalOrientation
@@ -69,8 +70,8 @@ class Explore:
     def disconnect(self):
         r"""Disconnects from the device
         """
-        self.stream_processor.stop()
         self.is_connected = False
+        self.stream_processor.stop()
         logger.debug("Device has been disconnected.")
 
     def acquire(self, duration=None):
@@ -136,7 +137,7 @@ class Explore:
         self.recorders['timer'].start()
         if block:
             try:
-                while self.recorders['timer'].is_alive():
+                while 'timer' in self.recorders.keys() and self.recorders['timer'].is_alive():
                     time.sleep(.3)
             except KeyboardInterrupt:
                 logger.info("Got Keyboard Interrupt while recording in blocked mode!")
@@ -245,7 +246,7 @@ class Explore:
 
         if block:
             try:
-                while self.lsl['timer'].is_alive():
+                while 'timer' in self.lsl.keys() and self.lsl['timer'].is_alive():
                     time.sleep(.3)
             except KeyboardInterrupt:
                 logger.info("Got Keyboard Interrupt while pushing data to LSL in blocked mode!")
@@ -349,23 +350,32 @@ class Explore:
     def set_channels(self, channel_mask):
         """Set the channel mask of the device
 
-        The channels can be disabled/enabled by calling this function and passing an integer which represents the
-        binary form of the mask. For example in a 4 channel device, if you want to disable channel 4, the adc mask
-        should be b'0111' (LSB is channel 1). The integer value of 0111 which is 7 must be given to this function
+        The channels can be disabled/enabled by calling this function and passing either bytes or binary string representing the mask.
+        For example in a 4 channel device, if you want to disable channel 4, the adc mask should be b'0111' (LSB is channel 1).
+        The inputs to this function can be b'0111' '0111'.
 
         Args:
-            channel_mask (int): Integer representation of the binary channel mask
+            channel_mask (bytes str): Bytes or String representing the binary channel mask
 
-        Examples:
+        Example:
             >>> from explorepy.explore import Explore
             >>> explore = Explore()
             >>> explore.connect(device_name='Explore_2FA2')
-            >>> explore.set_channels(channel_mask=7)  # disable channel 4 - mask:0111
+            >>> explore.set_channels(channel_mask='0111')  # disable channel 4 - mask:0111
         """
-        if not isinstance(channel_mask, int):
-            raise TypeError("Input must be an integer!")
+        c = re.compile('[^01]')
+
+        if (isinstance(channel_mask, str) and len(c.findall(channel_mask))==0) or (isinstance(channel_mask, bytes)):
+            channel_mask_int = int(channel_mask, 2)
+        # elif isinstance(channel_mask, int):
+        #     channel_mask_int = channel_mask
+        else:
+            # raise TypeError("Input must be an integer or a binary string!")
+            raise TypeError("Input must be bytes or binary string!")
+
         self._check_connection()
-        cmd = SetCh(channel_mask)
+        # cmd = SetCh(channel_mask)
+        cmd = SetCh(channel_mask_int)
         self.stream_processor.configure_device(cmd)
 
     def disable_module(self, module_name):